refactor: isolate cartographer utilities
